+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/absolute_target_speed.hpp
@@ -37,6 +37,23 @@ struct AbsoluteTargetSpeed
   : value{readAttribute<Double>("value", node, scope)}
   {
   }
+
+  auto getCalculateAbsoluteTargetSpeed() const
+  {
+    return [target_speed = value] { return target_speed; };
+  }
+
+  auto getIsEnd() const
+  {
+    return [target_speed = value](const Scope::Actor & actor) {  // is_end
+      try {
+        const auto compare = Rule(Rule::equalTo);
+        return compare(getEntityStatus(actor).action_status.twist.linear.x, target_speed);
+      } catch (const SemanticError &) {
+        return false;  // NOTE: The actor is maybe lane-changing now
+      }
+    };
+  }
 };
 }  // namespace syntax
 }  // namespace openscenario_interpreter