+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/storyboard.hpp
@@ -55,6 +55,10 @@ struct Storyboard : public StoryboardElement<Storyboard>, public Elements
     callWithElements(node, "Story", 1, unbounded, [&](auto && node) {
       return push_back(readStoryboardElement<Story>(node, inner_scope));
     });
+
+    if (not init.endsImmediately()) {
+      throw SemanticError("Actions in Init should end immediately.");
+    }
   }
 
   static constexpr auto ready() noexcept { return true; }