+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/user_defined_action.hpp
@@ -39,6 +39,8 @@ struct UserDefinedAction : public Element
       return rebind<CustomCommandAction>(node, scope);
     });
   }
+
+  static bool endsImmediately() { return true; }
 };
 }  // namespace syntax
 }  // namespace openscenario_interpreter